test(mobile-sync): smoke test AuthenticatedDevice extension on happy path (P5a.2)
middleware 早在 4dc83076(Phase 3 子步骤 5e/5f)就把 AuthenticatedDevice
塞进 request.extensions, 但当时没单测覆盖 —— P5a.3+ 业务 use case 要靠
这条契约从 axum Extension<AuthenticatedDevice> 拿来源 device_id, 缺测
就是个隐藏地雷。

新增 middleware::tests::happy_path_inserts_authenticated_device:
- 用最小 router (basic_auth + 探针 handler 提取 Extension<AuthenticatedDevice>)
- happy path 请求带 valid Basic Auth
- 断言 200 + body 是种子设备 username
- 同时验证 "extension 被插入" 与 "下游 handler 能取到" 两件事

顺手把 routes.rs 里 ~150 行 in-line fake-port fixture 抽到新的
mobile_lan::test_support (#[cfg(test)] 模块), middleware + routes 共用
同一份 build_facade_with_seeded_device + auth_header, 后续 P5a.3+ 的
单测也能直接复用。

更新 middleware.rs 末尾的 future-work 注释, 反映 P5a.3 会真正用 Extension
且这次已经有 smoke 用例覆盖契约。

Tests: cargo test -p uc-webserver mobile_lan -> 7 passed (原 6 + 新 1)。
